Extended the strictDecoding paradigm to BEncodedValues.
This allows for situations where one wants a torrent to decode no matter what - the default behavior is still the same (throw error).

I did this primarily because uTorrent 1.8 has a bug with torrent creation and setting keys out of order, and a lot of popular torrents have this problem. So, I made this fix so that they can still be decoded by MT.

git-svn-id: svn://anonsvn.mono-project.com/source/trunk/bitsharp@110296 e3ebcda4-bce8-0310-ba0a-eca2169e7518